+++ tests/testthat/test_setwrapper_productset.R
@@ -5,7 +5,7 @@ context("ProductSet")
 test_that("Interval",{
   useUnicode(FALSE)
   expect_equal((Interval$new(1,2) * ConditionalSet$new(function(x) TRUE))$strprint(),
-               "[1, 2] X {TRUE : x in R}")
+               "[1, 2] X {TRUE : x in V}")
   expect_equal(Interval$new(1,2)*Set$new(), Interval$new(1,2))
 
   i = (Interval$new(1, 2) * Interval$new(3, 4)) * Interval$new(5, 6)
@@ -27,14 +27,14 @@ test_that("Interval",{
 test_that("conditionalset",{
   useUnicode(FALSE)
   expect_equal((ConditionalSet$new(function(x) x == 1) * ConditionalSet$new(function(y) y > 1))$strprint(),
-               "{x == 1 : x in R} X {y > 1 : y in R}")
+               "{x == 1 : x in V} X {y > 1 : y in V}")
   useUnicode(TRUE)
 })
 
 test_that("fuzzy",{
   useUnicode(FALSE)
   expect_equal((FuzzySet$new(1,0.5) * ConditionalSet$new(function(x) TRUE))$strprint(),
-               "{1} X {TRUE : x in R}")
+               "{1} X {TRUE : x in V}")
   expect_true(setproduct(FuzzySet$new(1, 0.1), FuzzySet$new(2, 0.2), simplify = TRUE)$equals(Set$new(FuzzyTuple$new(1, 0.1, 2, 0.2))))
   expect_equal(setproduct(FuzzySet$new(1, 0.1), FuzzySet$new(2, 0.2), simplify = FALSE)$strprint(),
                "{1(0.1)} X {2(0.2)}")